server/requirements.txt
Django
djangorestframework
django-cors-headers
celery
redis>=3.5,<4.0
gunicorn
pydub
requests
feedparser
openai
python-dotenv
speechrecognition
pillow